Add an enabled field to queue - hide a queue without deleting it
If you decide a queue is not required or not suitable you have two choices;
- delete the queue, deleting all tickets on that queue
- edit every ticket, moving them to another queue

This change adds a feature where you can disable a queue. A disabled
queue wont show in the logged in or public create ticket form, but
can still be displayed when viewing a ticket from that disabled queue.