Re-enable status changes for non-staff users
I had originally requested not allowing these, but when reviewing the code
running, it's obvious that a non-staff user needs to be able to close
or re-open a ticket after an admin resolves it.

Given that, we might as well just let the user make any status changes
since anything else could be rather confusing.