Drop display of ticket priority
In my experience, this field is never useful. Every submitter
considers their own ticket a high priority, so there's no point in
asking them. And it's an even worse idea to have an admin lower the
priority of someone's ticket.

Much better to never display it.

While doing this, we drop the entire "Change Further Details"
section. All this allowed was changing owne and priority (which we
aren't using) and "Due on" (again, that's a kind of priority knob we
don't want to give to the user) and "title" which is just a chance to
rewrite history, (which is counter to the notion of tss).

So that whole section can disappear.