Drop the "Copies to" field from this view
From Todd's point of view, email is just a potential information link,
so we don't want a field that allows communication to go other than
where it was directly intended.

This makes the ticket system a fairly personal conversation between
ticket submitters and admins, (other than global tickets---which can
only be submitted by admins, so that's a broadcast mechanism from
admins to users).