Remove complexity from _aboslute_uri in models.py
Previously, the code would first try to build a URL using the request object
if one was available. However, the "if self.request" was often raising
an uncaught AttributeError because the request didn't exist. We address
this problem by simply always using the current Site to build the URL.
This has the added benefit of reducing the cyclomatic complexity of this
function, making the behavior easier to reason about and maintain.