General: Commented out user bugs
For some reason, SEED's app_label and model_name break Helpdesk when it tries to use them to reverse a url. (The result is something like landing_seeduser_[action].) In the interim, I removed the one url in the user settings page that requires this, as well as a method in staff.py that doesn't seem to need it yet.