DC CSS changes
Added DC logo to new images folder. Changed overall font and top navbar color. Moved account dropdown to the sidebar. In base.html, moved base_js.html to bottom of the page so that the sidebar toggle would work (as it does on the public side; not sure why original Helpdesk did it differently.