Perms: List of owners now includes roles >= member
The list of possible owners (found on the staff's ticket view and the Owner filter on the all-ticket list) was updated to only display users with roles of member and above. Needs to be updated to only display users in the current org. Also updated the lists to prefer displaying the user's full name over their email/username.