Migrations: Edited cascading on_delete & JSONField
Altered on_delete=models.CASCADE for several fields in Helpdesk to prevent deletion, especially when a user is deleted but their tickets or follow-up comments should remain. Also added a migration to use django.db.models.JSONField in place of django.contrib.postgres.fields.JSONField, to cut down on the warnings we get from that.