ticket list: added extra cols for ticket's extra data field
those new columns are only added when one queue is selected. since a two tickets can have two different forms with the same queue, they will have differente sets of extra_data. in the ticket list, that spot will be blank for the ticket that doesn't have that data. the filtering was also cleaned up and simplified