Public: Tickets can be accessed with multiple email fields
Public-side tickets can now be accessed through every kind of email field that could possibly be attached to a ticket. This includes:
- submitter email or contact email
- CC'd emails with view or edit permissions
- any extra_data email field with the type 'email' and with notifications on for that field
More changes are necessary to make this comprehensive (for example, the 'copy link' feature returns a link with the submitter's email, and I believe all Helpdesk-sent update emails also use the submitter_email). Eventually, it would be good for either submitter_email or contact_email to be used, instead of submitter_email always required...?