Email import: Added keep_mail field to queue
If a queue would like to keep the mail in their inbox, and not delete it after processing, check 'keep_mail'. This marks processed mail as 'answered', and ignores all mail that isn't marked 'answered'. Note that if multiple queues share an email address, they should all have the same setting set; in the future, move this field to the EmailSettings object for that email.