Email logging: Renamed ticket.send, added source & user args
Renamed the Ticket model's send() method to send_ticket_mail, to make it easier to search for instances of it in the code. Added the arguments source and user to send_templated_mail, and updated send_templated_mail and send_ticket_mail to include them where applicable. send_templated_mail adds the string "Helpdesk: " in front of the source before calling send_beam_mail.