Setting to add an list of bookmarks to the bottom of the sidebar menu.
A new setting HELPDESK_BOOKMARKS is added. Defaults to an empty list.

If it is set to to a list of dictionaries each list item is rendered
as either external or internal link in a separate section on the bottom
of the sidebar menu.

Bookmarks are specified as one of
- { 'URL': external_url, 'label': label } .. external link
- { 'name': internal_name, 'label': label } .. internal name, resolved with reverse()

If the 'label' is omitted, the URL or name is used as label.