Add JSON extra_fields field to followup module
This is extremely useful when writing custom frontends on top of django
helpdesk. This is similar to the custom_fields except it is normalized and
attached to followups and not tickets. The fact that it is just JSON means that
there is no need for updating the database with new CustomField entries making
for more easily maintained custom frontends.